Allow input of alternate definitions without specifying a type. This allows for convenient shorthand such as:
n. defn A/defn B

Both defns would have a type of noun.

Pretty large change. Had to change the dictionary a bit. Now the element
returned given a key, is a list of DictionaryEntrys. Where a
DictionaryEntry is simply a grouping of phrase type and phrase, e.g.
noun and 'hi'. Modify the code to handle this new dictionary type.

Seems to work mostly like before with 2 caveats:
1) You no longer need to input the phrase type, e.g. Noun. In fact, this
will be considered wrong if you do. I am not sure if I will bring back
this feature. It is kind of nice not having to type in phrase types.
I'll see how I feel about this later.
2) You can no longer enter in all the alternate definitions at once. As
there is no parsing of the user input. It is a straight up string
compare. I don't think it's a big deal right now. Getting one is
sufficient for most purposes. Will file a bug but will probably be lower
priority. There are a lot of other things to work on.